Make migration manga-centric rather than source-centric (#2786)
(cherry picked from commit cce3b3a559bc32b5866bfb0e294a9b45936b0f92)
