+++ app/src/main/java/eu/kanade/tachiyomi/ui/migration/MigrationController.kt
@@ -1,19 +1,14 @@
 package eu.kanade.tachiyomi.ui.migration
 
-import android.app.Dialog
-import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.afollestad.materialdialogs.MaterialDialog
+import androidx.recyclerview.widget.LinearLayoutManager
 import eu.davidea.flexibleadapter.FlexibleAdapter
 import eu.davidea.flexibleadapter.items.IFlexible
 import eu.kanade.tachiyomi.R
 import eu.kanade.tachiyomi.data.database.DatabaseHelper
-import eu.kanade.tachiyomi.data.database.models.Manga
-import eu.kanade.tachiyomi.ui.base.controller.DialogController
 import eu.kanade.tachiyomi.ui.base.controller.NucleusController
-import eu.kanade.tachiyomi.ui.base.controller.popControllerWithTag
 import eu.kanade.tachiyomi.ui.base.controller.withFadeTransaction
 import exh.ui.migration.manga.design.MigrationDesignController
 import exh.util.await
@@ -51,7 +46,7 @@ class MigrationController : NucleusController<MigrationPresenter>(),
         super.onViewCreated(view)
 
         adapter = FlexibleAdapter(null, this)
-        migration_recycler.layoutManager = androidx.recyclerview.widget.LinearLayoutManager(view.context)
+        migration_recycler.layoutManager = LinearLayoutManager(view.context)
         migration_recycler.adapter = adapter
     }
 
@@ -91,16 +86,6 @@ class MigrationController : NucleusController<MigrationPresenter>(),
         }
     }
 
-    fun renderIsReplacingManga(state: ViewState) {
-        if (state.isReplacingManga) {
-            if (router.getControllerWithTag(LOADING_DIALOG_TAG) == null) {
-                LoadingController().showDialog(router, LOADING_DIALOG_TAG)
-            }
-        } else {
-            router.popControllerWithTag(LOADING_DIALOG_TAG)
-        }
-    }
-
     override fun onItemClick(view: View?, position: Int): Boolean {
         val item = adapter?.getItem(position) ?: return false
 
@@ -129,28 +114,5 @@ class MigrationController : NucleusController<MigrationPresenter>(),
                 router.pushController(MigrationDesignController.create(sourceMangas).withFadeTransaction())
             }
         }
-    }
-
-    fun migrateManga(prevManga: Manga, manga: Manga) {
-        presenter.migrateManga(prevManga, manga, replace = true)
-    }
-
-    fun copyManga(prevManga: Manga, manga: Manga) {
-        presenter.migrateManga(prevManga, manga, replace = false)
-    }
-
-    class LoadingController : DialogController() {
-
-        override fun onCreateDialog(savedViewState: Bundle?): Dialog {
-            return MaterialDialog.Builder(activity!!)
-                    .progress(true, 0)
-                    .content(R.string.migrating)
-                    .cancelable(false)
-                    .build()
-        }
-    }
-
-    companion object {
-        const val LOADING_DIALOG_TAG = "LoadingDialog"
     }
 }
