+++ app/src/main/java/eu/kanade/tachiyomi/ui/migration/ViewState.kt
@@ -5,6 +5,5 @@ import eu.kanade.tachiyomi.source.Source
 data class ViewState(
     val selectedSource: Source? = null,
     val mangaForSource: List<MangaItem> = emptyList(),
-    val sourcesWithManga: List<SourceItem> = emptyList(),
-    val isReplacingManga: Boolean = false
+    val sourcesWithManga: List<SourceItem> = emptyList()
 )
